feat: reusable sort by field
